requirements.txt
backcall==0.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
Pygments==2.4.2
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
tqdm==4.38.0
traitlets==4.3.3
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7